Implement GSUP support
Osmocom's [GSUP][1] is a minimalistic protocol that is used by their 2G
components (OsmoMSC, OsmoSGSN) in place of MAP/TCAP as it
is a lot simpler.

At [C3GSM][2], we are operating our own 2G/3G/4G networks during Chaos
Computer Club events like the recent 38C3 conference. Currently, the 4G
network is integrated using a Diameter to GSUP protocol proxy, that
allows us to use OsmoHLR for 4G, but the only implemented procedures are
for Authentication and UpdateLocation.

In order to facilitate IMS / VoLTE / VoWifi in the future, it would be
convenient to replace OsmoHLR with PyHSS but also retain the 2G/3G stack
of OsmoSGSN and OsmoMSC. Therefore, we implemented the GSUP protocol and
2G/3G procedures into PyHSS, enabling an open source 2G/3G/4G network
with one central brain.

Implemented procedures are:

- sendAuthInfo
- updateLocation
- insertSubscriber
- cancelLocation
- purgeMS

This implementation is working and has been tested during 38C3.
We'd like to contribute our changes back upstream as this might help
other people to also run their own 2G/3G/4G combined networks and use
Osmocom for their 2G/3G needs.

[1]: https://ftp.osmocom.org/docs/osmo-hlr/master/osmohlr-usermanual.pdf
[2]: https://c3gsm.de/

Co-Authored-By: Alexander Couzens <lynxis@fe80.eu>